backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
PyJWT
httpx
anthropic
azure-identity
beautifulsoup4
python-multipart
PyPDF2
python-docx
